backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
rembg==2.0.50
gunicorn==21.2.0
boto3==1.28.62
python-dotenv==1.0.0
scikit-learn==1.3.0 